Moved the favorite listings all into the same ocmponent for testing.
