+++ src/components/Profile/Profile.js
@@ -3,13 +3,7 @@ import { makeStyles } from '@material-ui/core/styles';
 import Paper from '@material-ui/core/Paper';
 import Avatar from '@material-ui/core/Avatar';
 import Rating from '@material-ui/lab/Rating';
-import Grid from '@material-ui/core/Grid';
-import Box from '@material-ui/core/Box';
-import Typography from '@material-ui/core/Typography';
-import Skeleton from '@material-ui/lab/Skeleton';
-import IconButton from '@material-ui/core/IconButton';
-import FavoriteIcon from '@material-ui/icons/Favorite';
-import ShareIcon from '@material-ui/icons/Share';
+import Favorites from './Favorites';
 
 const useStyles = makeStyles(theme => ({
   root: {
@@ -39,89 +33,12 @@ const useStyles = makeStyles(theme => ({
     backgroundSize: 'cover',
     backgroundPosition: 'center',
   },
-  example: {
-    display: 'flex',
-    flexDirection: 'column',
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: 'lightblue',
-    margin: '1% auto 0',
-    width: '80%',
-    height: '425px',
-    padding: '1%'
-  },
-  grid: {
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-    margin: '0 auto',
-    width: '80%',
-    padding: '0 1%',
-    cursor: 'pointer',
-  },
-  favContent: {
-    display: 'flex',
-    flexDirection: 'column',
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#fff',
-    margin: '1% auto',
-    width: '100%',
-    padding: '1%',
-    boxShadow: '1px 3px 5px'
-  },
-  heading: {
-    fontSize:'1.6rem',
-    margin: '0 auto'
-  },
-  image: {
-    '&:hover': {
-      transform: 'scale(1.05)',
-    }
-  }
 }));
 
-const data = [
-  {
-    src: 'https://source.unsplash.com/random',
-    title: 'Property One',
-    location: 'San Jose, CA',
-    views: '396 k views',
-    createdAt: 'a week ago',
-  },
-  {
-    src: 'https://source.unsplash.com/random',
-    title: 'Property Two',
-    location: 'Paris, France',
-    views: '2.1 M views',
-    createdAt: '4 months ago',
-  },
-  {
-    src: 'https://source.unsplash.com/random',
-    title: 'Property Three',
-    location: 'Tampa, FL',
-    views: '130 M views',
-    createdAt: '10 months ago',
-  },
-  {
-    src: 'https://source.unsplash.com/random',
-    title: 'Property One',
-    location: 'San Jose, CA',
-    views: '396 k views',
-    createdAt: 'a week ago',
-  },
-  {
-    src: 'https://source.unsplash.com/random',
-    title: 'Property Two',
-    location: 'Paris, France',
-    views: '2.1 M views',
-    createdAt: '4 months ago',
-  },
-];
 
-function Profile(props) {
+
+function Profile() {
   const classes = useStyles();
-  const { loading = false } = props;
 
   return (
     <>
@@ -135,49 +52,8 @@ function Profile(props) {
         <Rating name="half-rating" value={4.5} precision={0.5} />
       </Paper>
     </Paper>
-    <Paper className={classes.example}>
-    <h1 className={classes.heading}>Favorite Properties</h1>
-    <Grid container wrap="nowrap" className={classes.grid}>
-    {(loading ? Array.from(new Array(3)) : data).map((item, index) => (
-      <Box key={index} width={260} margin={1} my={5} className={classes.image}>
-        {item ? (
-          <img style={{ width: 220, height: 140 }} alt={item.title} src={item.src} />
-        ) : (
-          <Skeleton variant="rect" width={260} height={140} />
-        )}
-
-        {item ? (
-          <Box paddingRight={2} className={classes.favContent}>
-            <Typography gutterBottom variant="body2">
-              {item.title}
-            </Typography>
-            <Typography display="block" variant="caption" color="textSecondary">
-              {item.location}
-            </Typography>
-            <Typography variant="caption" color="textSecondary">
-              {`${item.views} • ${item.createdAt}`}
-            </Typography>
-            <Rating name="half-rating" value={4.5} precision={0.5} />
-            <div>
-              <IconButton aria-label="add to favorites">
-                <FavoriteIcon />
-              </IconButton>
-              <IconButton aria-label="share">
-                <ShareIcon />
-              </IconButton>
-            </div>
-          </Box>
-        ) : (
-          <React.Fragment>
-            <Skeleton />
-            <Skeleton width="60%" />
-          </React.Fragment>
-        )}
-      </Box>
-      ))}
-    </Grid>
-  </Paper>
-  </>
+    <Favorites />
+    </>
   )
 }
